Added PickUpController Script. Blasters can now be picked up and dropped.
Add the pickUpController Script to any Gun that you want to be pickable and dropped. Use "E" to pick up and "G" to Drop.
Currently no way to specify the position you want the Gun snap to when picked. I will add that functionality with the inventory system. And also make it more general so any object can be pickable, not just Guns.