Implemented Inventory system and some UI changes
The player object has a new script called player Inventory. Any Interactable object/class  inheriting from the IIinvetoryItems class can be added to the inventory. If an item is consumable/stackable then the Objectname of that consumable must be the same.
The ammo Text UI now changes color based on element.